// spending-tracker-frontend/src/interfaces/Entry.tsx
export interface Entry {
    date: string;
    income: number;
    startOfDayBalance: number;
    coldCash: number;
    grocery: number;
    fastFood: number;
    bills: number;
    subscriptions: number;
    gas: number;
    shopping: number;
    miscellaneous: number;
    robinHoodTransfer: number;
    endOfDayBalance: number;
    robinHood: number;
    totalAssets: number;
    percentChange: number;
}